Change estimation of the object's orientation due to the new camera setting.
In the past, the camera was placed on the left of the robot, so its Z axe was pointing towards the table and the Y axe was pointing down. Now, the camera is at the top, with the Z axe pointing down towards the table. Thus, we now have to use X and Y axis for finding the initial points, instead of the X and Z axis.