Add custom grasp configuration message and update the grasp computation for the camera on top case.
We are using in this branch a camera whose point of view is perpendicular to the table, being it on top of it. Thus, some minor changes have been made to the GeoGrasp computation. Besides, a custom message has been created for publishing grasps, so another node can calculate a grasp pose from this data as explained in the IJARS paper.